Prevent accidentally deleting a dataset by adding a confirmation dialog
Can later be reworked to be a modal window that fits in with the rest of the elements on the page instead of the default browser dialog alert

+++ client/src/components/DatasetList.js
@@ -57,9 +57,11 @@ class DatasetList extends React.Component {
 
   handleDelete = (id) => {
     const { deleteDataset, navigation } = this.props
-    deleteDataset({ variables: { id }, refetchQueries: [{ query: datasetListQuery }] })
-    if (id === navigation.currentDataset) {
-      navigation.selectDataset(null)
+    if (confirm("Are you sure you want to remove this dataset?")) {
+      deleteDataset({ variables: { id }, refetchQueries: [{ query: datasetListQuery }] })
+      if (id === navigation.currentDataset) {
+        navigation.selectDataset(null)
+      }
     }
   }
